// src/app/core/interfaces/cedula-presupuestal-detalle-form.interface.ts
export interface CedulaPresupuestalDetalleForm {
  id: any;
  cedulaPresupuestalid: any;
  cuentaid: any;
  presupuestoMensual: number;
  presupuestoEjercido: number;
  presupuestoAcumulado: number;
  presupuestoDisponible: number;
  presupuestoAnual: number;
  presupuestoRestanteAnio: number;
  applicationUserId: string;
}